Override transitive base64 dependency
It seems like with aws-sdk-ec2 1.515.0, the version of base64 was
increased to 0.2.0. But one of the transitive dependencies of this gem
wants 0.1.1, so we couldn't install the plugin.

We tried specifying version 0.2.0 of `base64` in the `.gemspec` file,
but still got the same error. So I went to Google to understand how to
override transitive dependencies.

I found from a Github issue opened Sept 2024:
https://github.com/rubygems/rubygems/issues/8021

Where the author had created a Bundler plugin, bundler-resolutions:
https://github.com/hlascelles/bundler-resolutions

And this does indeed solve the problem! However, installing some brand
new experimental plugin is probably not how people normally solve
problems. There is almost certainly a much, much better way to do this
that would be totally clear to someone who understands the Ruby
ecosystem. Unfortunately, our team doesn't use Ruby, so I'm a bit
clueless on the subject.

Anyway, I'm pushing this in the hopes of getting help from experienced
Rubyists, while unblocking our team in the meantime :)